created taskEdit.component.html for dialog task edits, updated task component html, ts, for task dialogs. changed several task.name to task.title to fix and have consisitent variable name. changed name to task within interface as well.

+++ client/src/app/components/list/list.component.ts
@@ -70,20 +70,24 @@ export class ListComponent {
 
   openEditList(): void {
     const dialogRef = this.dialog.open(listEditComponent, {
-      data: { lists: this.lists, name: this.newListName }, height:'425px',
+      data: { lists: this.lists, name: this.newListName }, height: '425px', width: '250px'
     });
 
     dialogRef.afterClosed().subscribe(result => {
-      if (this.lists.some(list => list._id === result)) { // If result exists in lists, it's a delete action
-        this.listTaskService.deleteList(result).subscribe(() => {
-          this.fetchLists();
-        });
-      } else { // Otherwise, new list creation
-        this.listTaskService.createList({ name: result }).subscribe(() => {
-          this.fetchLists();
-        });
+      if (result) {
+        if (this.lists.some(list => list._id === result)) { // If result exists in lists, it's a delete action
+          this.listTaskService.deleteList(result).subscribe(() => {
+            this.fetchLists();
+          });
+        } else { // Otherwise, new list creation
+          this.listTaskService.createList({ name: result }).subscribe(() => {
+            this.fetchLists();
+          });
+        }
+      } else {
+        console.log('Edit List Cancelled');
+        
       }
-
     });
   }
 }
